Remove DummyBlock
