+++ crates/types/src/traits/block_contents.rs
@@ -76,125 +76,3 @@ pub trait BlockHeader:
     /// Get the payload commitment.
     fn payload_commitment(&self) -> Commitment<Self::Payload>;
 }
-
-/// Dummy implementation of `BlockPayload` for unit tests
-pub mod dummy {
-    use std::fmt::Display;
-
-    use super::{
-        BlockHeader, BlockPayload, Commitment, Committable, Debug, Hash, HashSet, Serialize,
-    };
-    use rand::Rng;
-    use serde::Deserialize;
-
-    pub use crate::traits::state::dummy::DummyState;
-    use crate::traits::state::TestableBlock;
-
-    // TODO (Keyao) Investigate the use of DummyBlock.
-    // <https://github.com/EspressoSystems/HotShot/issues/1763>
-    /// The dummy block
-    #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
-    pub struct DummyBlock {
-        /// Some dummy data
-        pub nonce: u64,
-    }
-
-    impl DummyBlock {
-        /// Generate a random `DummyBlock`
-        pub fn random(rng: &mut dyn rand::RngCore) -> Self {
-            Self { nonce: rng.gen() }
-        }
-    }
-
-    /// Dummy error
-    #[derive(Debug)]
-    pub struct DummyError;
-
-    /// dummy transaction. No functionality
-    #[derive(Debug, Clone, Copy, Eq, PartialEq, Serialize, Deserialize, Hash)]
-    pub enum DummyTransaction {
-        /// the only variant. Dummy.
-        Dummy,
-    }
-
-    impl Committable for DummyTransaction {
-        fn commit(&self) -> commit::Commitment<Self> {
-            commit::RawCommitmentBuilder::new("Dummy BlockPayload Comm")
-                .u64_field("Dummy Field", 0)
-                .finalize()
-        }
-
-        fn tag() -> String {
-            "DUMMY_TXN".to_string()
-        }
-    }
-    impl super::Transaction for DummyTransaction {}
-
-    impl std::error::Error for DummyError {}
-
-    impl std::fmt::Display for DummyError {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            f.write_str("A bad thing happened")
-        }
-    }
-
-    impl Display for DummyBlock {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{self:#?}")
-        }
-    }
-
-    impl BlockHeader for DummyBlock {
-        type Payload = Self;
-
-        fn new(_payload_commitment: Commitment<Self>, _parent_header: &Self) -> Self {
-            Self { nonce: 0 }
-        }
-
-        fn genesis(_payload: Self::Payload) -> Self {
-            Self { nonce: 0 }
-        }
-
-        fn block_number(&self) -> u64 {
-            0
-        }
-
-        fn payload_commitment(&self) -> commit::Commitment<Self> {
-            commit::RawCommitmentBuilder::new("Dummy BlockPayload Comm")
-                .u64_field("Nonce", self.nonce)
-                .finalize()
-        }
-    }
-
-    impl BlockPayload for DummyBlock {
-        type Error = DummyError;
-
-        type Transaction = DummyTransaction;
-
-        fn transaction_commitments(&self) -> HashSet<Commitment<Self::Transaction>> {
-            HashSet::new()
-        }
-    }
-
-    impl TestableBlock for DummyBlock {
-        fn genesis() -> Self {
-            Self { nonce: 0 }
-        }
-
-        fn txn_count(&self) -> u64 {
-            1
-        }
-    }
-
-    impl Committable for DummyBlock {
-        fn commit(&self) -> commit::Commitment<Self> {
-            commit::RawCommitmentBuilder::new("Dummy BlockPayload Comm")
-                .u64_field("Nonce", self.nonce)
-                .finalize()
-        }
-
-        fn tag() -> String {
-            "DUMMY_BLOCK".to_string()
-        }
-    }
-}
